Fix lint issues.
